# openapi.yaml
openapi: "3.0.3"
info:
  title: fortbyte
  version: "0.1.0"
  description: |
    Personal secrets manager with a hosted REST API.
    Supports JWT access tokens, refresh token rotation, and API key authentication.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      operationId: refresh
      summary: Refresh tokens
      description: Exchanges a refresh token for a new access + refresh token pair (rotation).
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: The refresh token to exchange
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Use a JWT access token (returned from register/login) or an API key prefix `fb_`.
        JWT tokens expire after 15 minutes; refresh via `/auth/refresh`.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (15 min TTL)
        refresh_token:
          type: string
          description: Refresh token (7 day TTL)
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 900
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: auth_error
            message:
              type: string
              example: invalid credentials